+++ packages/cli/tests/unit/ui.tests.ts
@@ -483,168 +483,3 @@ describe("User Interaction Tests", function () {
     UI.createProgressBar("title", 3);
   });
 });
-
-describe("Errors in User Interaction", async () => {
-  const sandbox = sinon.createSandbox();
-  afterEach(() => {
-    sandbox.restore();
-  });
-  it("InputValidationError", async () => {
-    const config: InputTextConfig = {
-      name: "testInput",
-      title: "input text",
-      validation: (input: string) => {
-        return "failed";
-      },
-    };
-    UI.updatePresetAnswer("testInput", "valuebrabrabra");
-    const result = await UI.inputText(config);
-    expect(result.isErr());
-    if (result.isErr()) {
-      expect(result.error.name).equals("InputValidationError");
-    }
-  });
-
-  it("UnhandledError", async () => {
-    sandbox.stub(inquirer, "prompt").rejects(new Error("test"));
-    const config: InputTextConfig = {
-      name: "testInput",
-      title: "input text",
-    };
-    UI.clearPresetAnswers();
-    const result = await UI.inputText(config);
-    expect(result.isErr());
-    if (result.isErr()) {
-      expect(result.error.name).equals("UnhandledError");
-    }
-  });
-
-  it("SelectSubscriptionError", async () => {
-    sandbox.stub(inquirer, "prompt").rejects(new Error("test"));
-    const config: SingleSelectConfig = {
-      name: "subscription",
-      title: "select subscription",
-      options: [],
-    };
-    const result = await UI.selectOption(config);
-    expect(result.isErr());
-    if (result.isErr()) {
-      expect(result.error.name).equals("SelectSubscriptionError");
-    }
-  });
-});
-
-describe("User Interaction Tests for select", function () {
-  const sandbox = sinon.createSandbox();
-  beforeEach(() => {
-    sandbox.stub(UI, "createProgressBar").returns({
-      start: async (s) => {},
-      next: async (s) => {},
-      end: async (s) => {},
-    });
-  });
-  afterEach(() => {
-    sandbox.restore();
-  });
-  describe("loadSelectDynamicData", async () => {
-    it("happy path", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: async () => ["a", "b", "c"],
-        default: async () => "a",
-      };
-      const result = await UI.loadSelectDynamicData(config);
-      expect(result.isOk());
-      expect(config.options).deep.equals(["a", "b", "c"]);
-      expect(config.default).equals("a");
-    });
-    it("throw error", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: async () => {
-          throw new Error("test");
-        },
-      };
-      const result = await UI.loadSelectDynamicData(config);
-      expect(result.isErr());
-    });
-    it("no need to call function", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: ["a", "b", "c"],
-      };
-      const result = await UI.loadSelectDynamicData(config);
-      expect(result.isOk());
-      expect(config.options).deep.equals(["a", "b", "c"]);
-    });
-  });
-
-  describe("loadDefaultValue", async () => {
-    it("happy path", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: ["a", "b", "c"],
-        default: async () => "a",
-      };
-      const result = await UI.loadDefaultValue(config);
-      expect(result.isOk());
-      expect(config.default).equals("a");
-    });
-    it("throw error", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: ["a", "b", "c"],
-        default: async () => {
-          throw new Error("test");
-        },
-      };
-      const result = await UI.loadDefaultValue(config);
-      expect(result.isErr());
-    });
-    it("no need to call function", async () => {
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: ["a", "b", "c"],
-        default: "a",
-      };
-      const result = await UI.loadDefaultValue(config);
-      expect(result.isOk());
-    });
-  });
-
-  describe("selectOptions dynamic data", async () => {
-    it("throw error", async () => {
-      sandbox.stub(UI, "loadSelectDynamicData").resolves(err(new UserError({})));
-      const config: MultiSelectConfig = {
-        name: "test",
-        title: "test",
-        options: async () => {
-          throw new Error("test");
-        },
-      };
-      const result = await UI.selectOptions(config);
-      expect(result.isErr());
-    });
-  });
-
-  describe("selectOption dynamic data", async () => {
-    it("throw error", async () => {
-      sandbox.stub(UI, "loadSelectDynamicData").resolves(err(new UserError({})));
-      const config: SingleSelectConfig = {
-        name: "test",
-        title: "test",
-        options: async () => {
-          throw new Error("test");
-        },
-      };
-      const result = await UI.selectOption(config);
-      expect(result.isErr());
-    });
-  });
-});
